Add test fixture for image resizing
In the v1 image resource integration test they altered the code to
reflect the case that a given image is resized. But the test code
is generated by JHipster. So we should change as less as possible
in this generated test files.

This patch adds the givne and expected image bytes as text fixtures
and exposes them as static method. This makes it possible to add this
test data in the generated test by only changeing a small part of a
single line.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>