Use the text fixture for images in image resource integration test
Sinde the image service does some byte mangling on images the generated
byte array is not sufficient as test fixture because it results in a NPE.

This patch adds the test fixture from v1 as an external helper in the test.
The reason forthat is to change the generated test class as less as
possible.

Some bytes does not match and thus the test is temporarily disabled.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>