Refactor the audit converter
This patch refactors the conversion methods to persistent events
that way that it utilizes streams API completely.

The former implementation was very convoluted, over complicated
and  not very efficient: iterating a given map over and over.

Now with streaming API the given data map is iterated only once
and the transformation methods are way more simpler.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>