Remove unwanted persitent audit event resources
We use the persitent audit events only internal to store them into
the databse. These are used in custom implementations of Spring
Boot interfaces for audit events. So these resources are not neccessery.

It is not a good practice to have code in the project which is not needed.
So this patch removes these files and ignores them, so they can not be
commited accidentally.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>