Fix one of the image resource integration tests
Problem was that tere was never triggered proper resizing if there is
large imput image and then the results to not match the expected files.

Another problem (not fixed yet) is that resizing is always triggered,
even if there are small and medium sizes. The result is that the already
resized large one is resizedagain. Which will end in slightly different
small and medium ones.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>