Rework the repository for achievable skills
What are achievable skills? AFAICS it is a join over teams and skills.
It has no entity in the JDL and therefor no reposiotory. As a workaround
in v1 they just added the queries to the SkillReposiotry.

As we followed the pattern to not alter generated code we put these queries
into an extending reposiotry.

This has some flaws from an architectural point of view:

- It extyends and therefore couples to an unrelated repository.
- Since there is no definition in the JDL there is also no entity
  and so a handcraftet DTO was used in the repository methods. But
  DTOs should not be used between services and reposiotries in this
  architecture.

This patch decouples the repository for achievable skills into a
stand-alone Spring repository.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>