Introduce entity and mapper for achievable skills
The whole application uses an architectural pattern to distinguish
entities to transfer data from reposiories to the services and DTOs
to transfer data between services and contollers.

Therefore entities must not be used in contollers and DTOs must
not be used in repositories.

The solution is to introduce a synthetic domain entity which is
used for the repository and a mapper which translates from/to the DTO.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>