Add teams.module and teams-achievement component
This patch adds only those parts of the components, which
didn't require the addition of other dependencies.

Moving the V1 code into V2 resulted in several linting errors.
This was caused by missing undefined and null checks inside
the V1 code.
The reason for this are certain entity fields in V2 are nullable,
which wasn't the case in V1.

To not push a non compiling state the changes were immediately
applied. To see a diff between the v1 and v2 version see
inside contrib folder.