Add //@ts-ignore to suppress linting error
After migrating the level-util.ts file, the linter throws
following error: "TS2345: Argument of type 'ILevel | undefined' is
not assignable to parameter of type 'ILevel'.   Type 'undefined'
is not assignable to type 'ILevel'."

This seems to arise from calling the pop() function on levels.
However the documentation states that undefined is only returned
if the array is empty. This won't happen in this case, as the
length of the levels is checked to be greater then 0 before
calling pop().