Fix type issues by changing IAchievableSkill to ISkill
onSkillClocked emits an object containing two fields one of which
is of type IAchiveableSkill. However the only place, which emits
these values assigns a variable of type ISkill to this field.

From looking at the V1 history it is not clear why this field
is assigned in such a way.