#42 Implement tracking of creation and update time fordomain model
We need to ensure that each entity which has an createdAt and
updatedAt field get the right values in there. The basic idea
is described in ADR-0002.

It is impleented by encapsualting the logic of determining
andsetingthese two fields into the helper class
AuditableDataTracker. To minimize coupling theclass is implemented
generic and the required API of the tracked DTOs isextracted
into the interface AuditableData. This interface is implemented in
the DTOs with absolute class name and no import to reduce the manual
changes in this generated class.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>