Deliberately ignore tests on generated DTOs
Since we removed the not-null validation for createdAt and updatedAt
in the DTOs these tests will always fail. We removed this validation
because we do not want these values set by the client, but instead
by the backend services.

Why not simply delete these tests? These test files are generated by
JHipster. So, we want to reduce manual changes to these fails to a
minimum. Removeing the whole test method is way more change than the
annotation for disabling. Also we can add our modification eye catcher.
This wouldn't be possible if we just remove the method.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>