#42 inject static instant provider in resource integration tests
Since we set the createdAt and updatedAt values for each incomming
DTO before it is persisted as entity, these tests fail which assume
that the value passed in by the HTTP request will be persisted.

These tes tfailures are fixed by injecting a time provider instance
to the underlying services which always return a fixed time which we
can assert.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>