#58 Remove obsolete regression test
Test is obsolete bc it only shows the system works as designed:
If a DTO w/o id is given as transitive entity, JPA complains
that it is not saved.

The bug in this case is in the frontend: You can add a comment w/o
selecting a team beforehand. It must not be possible to add a
comment w/o selecting a team.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>