#82 Removes custom TeamsService
In fact the custom teams.service has the same implementation as the
generated team.service.

This introduced the bug because the converter method didn't respect
the that datetime strings must be converted to Dayjs objects.

This fixes the issue because the generated service does this conversion.
Using the generated serice is more resillient because it will be
changed on code genertion if we add/remove properties which represent
datetime.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>