#5 Port team group to component
Using Angular's resolve is not sufficient because it only fetches the
data once at setup time of the nav bar. Since we use more than one group,
instead of one global orgfanization, this is not sufficient anymore.

This is an attempt to resolve this by adding a own component which
subscribes on the team selection service.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>