#8 Implement dynamic translation for model classes in the frontend
Basic idea:

We have almost always two properties (title & description) which are
localized in the database. This is reflected by adding the language
to the respective property name (titleDE, titleEN, descriptionDE,
descriptionEN).

We provide a unified interface with properties title and description.
On accesing these properties we dispatch dynamically to the right
loclaized property for the current language. This language is  simply
determined  by a service.

Implementation:

We provide an interface ITranslated to define the two properties (title
and description) and a implementation with dynamic dispatching. Also
we provide a stub language service which returns EN hardcoded at the
moment.

Skill is a special case because the model introduces two more localized
properties: implementation and validation.

Next steps:

1. Extend all the loclaized models.
2. Only use .title and .description in the frontend components instead
   of titleEN and descriptionEN.
3. Implement the language service to determine user's language.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>