#8 Decorator approach to translate user generated content
Since we can not use services in models and modifying the generated
models is a bad idea anyway, I decided to use decorator pattern
to add the functionality to the models.

This patch moves the whole logic into decorators. The client code
which wants to translate the data need to wrap the origin model
via factory (e.g. in the component). In the component templates
the original properties (e.g. title, description etc.) provide
the translated content depending on selected language.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>