Issue #126: Display current teamgroup title, when editing team
To display a preselected option the ngModel property compares
primitive values. Binding ngModel to a variable of non
primitive data type leads to the compairson failing,
which in turn results in no preselected option being displayed.
This can be fixed by using the compareWith property to
introduce a custom comparator function.