Issue #144: Add python script to Migrate class instanciation
In Jhipster 7.9.x entites, in the frontend, are only present as
interfaces. This changes the way in which those entites are
instanciated. The added python script is supposed to automate the
process of migrating entity instanciations in custom code files.

Besides the script dummy files are added for testing purposes.